AutoTestClient/requirements.txt
openpyxl==3.0.7
Flask==1.1.2
pywifi==1.1.12
psutil==5.8.0
pysmb==1.2.7
requests==2.24.0
selenium==3.141.0
PyYAML==5.4.1
comtypes==1.1.10